requirements.txt
pandas>=2.0
numpy>=1.24
psycopg2-binary>=2.9
anthropic>=0.18
